requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
sqlalchemy
psycopg2-binary
groq
